WIP - keep TS config tests tegether
